refactor(api,datamodel): aligned module names and documentation
