+++ libs/datamodel/api/README.md
@@ -1,7 +1,7 @@
-# shared-datamodel-api
+# datamodel-api
 
 This library was generated with [Nx](https://nx.dev).
 
 ## Running unit tests
 
-Run `ng test shared-datamodel-api` to execute the unit tests via [Jest](https://jestjs.io).
+Run `ng test datamodel-api` to execute the unit tests via [Jest](https://jestjs.io).
