+++ libs/datamodel/common/README.md
@@ -1,7 +1,7 @@
-# shared-datamodel
+# datamodel-common
 
 This library was generated with [Nx](https://nx.dev).
 
 ## Running unit tests
 
-Run `ng test shared-datamodel` to execute the unit tests via [Jest](https://jestjs.io).
+Run `ng test datamodel-common` to execute the unit tests via [Jest](https://jestjs.io).
